+++ seacatauth/authz/utils.py
@@ -15,16 +15,18 @@ async def build_credentials_authz(tenant_service, role_service, credentials_id, 
 	"""
 
 	# Add global roles and resources under "*"
-	authz = {"*": set()}
-	for role in await role_service.get_roles_by_credentials(credentials_id, "*"):
-		authz["*"].update(await role_service.get_role_resources(role))
-	authz["*"] = list(authz["*"])
+	authz = {}
+	tenant = "*"
+	authz[tenant] = set()
+	for role in await role_service.get_roles_by_credentials(credentials_id, [tenant]):
+		authz[tenant].update(await role_service.get_role_resources(role))
+	authz[tenant] = list(authz[tenant])
 
 	# Add tenant-specific roles and resources
 	if tenant_service.is_enabled() and tenants is not None:
 		for tenant in tenants:
 			authz[tenant] = set()
-			for role in await role_service.get_roles_by_credentials(credentials_id, tenant):
+			for role in await role_service.get_roles_by_credentials(credentials_id, [tenant]):
 				authz[tenant].update(await role_service.get_role_resources(role))
 			authz[tenant] = list(authz[tenant])